pyproject.toml
[project]
name = "dhis2-era5land"
version = "0.1.0"
description = "Import ERA5-Land climate data into DHIS2"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dhis2-client",
    "dhis2eo",
    "earthkit>=0.13.2",
    "fastapi>=0.115.0",
    "pydantic-settings>=2.7.0",
    "typer>=0.21.0",
    "uvicorn>=0.34.0",
]

[project.scripts]
dhis2-era5land = "dhis2_era5land.cli:app"

[dependency-groups]
dev = [
    "httpx>=0.28.0",
    "mkdocs>=1.6.0",
    "mkdocs-material>=9.5.0",
    "mypy>=1.18.2",
    "pyright>=1.1.407",
    "pytest>=8.0.0",
    "ruff>=0.14.10",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/dhis2_era5land"]

[tool.uv.sources]
dhis2eo = { git = "https://github.com/dhis2/dhis2eo.git" }
dhis2-client = { git = "https://github.com/dhis2/dhis2-python-client.git" }

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
fixable = ["ALL"]
select = ["E", "W", "F", "I", "D"]
ignore = ["D203", "D213"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["D"]
"**/__init__.py" = ["D104"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
docstring-code-format = true
docstring-code-line-length = "dynamic"

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_unused_ignores = true
strict_equality = true
mypy_path = ["src"]

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

[[tool.mypy.overrides]]
module = [
    "geopandas",
    "dhis2_client",
    "dhis2eo.*",
    "earthkit.*",
]
ignore_missing_imports = true

[tool.pyright]
include = ["src"]
pythonVersion = "3.12"
typeCheckingMode = "standard"
reportMissingTypeStubs = false
reportUnknownMemberType = false
reportUnknownArgumentType = false
reportUnknownParameterType = false
reportUnknownVariableType = false
reportMissingParameterType = false
